refactor spin generator
